refactoring to component files for clarity
